compiles_with_flags_now

+++ env/env_contain.c
@@ -6,7 +6,7 @@
 /*   By: david <david@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/10 18:33:54 by david             #+#    #+#             */
-/*   Updated: 2022/02/10 18:49:11 by david            ###   ########.fr       */
+/*   Updated: 2022/02/11 14:34:16 by david            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	env_contain(char *name, t_env *env)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (!name || !env || !env->vars)
